:recycle: Oppdatert tag-demo
