--- aksel.nav.no/website/pages/eksempler/tag/alt3.tsx
@@ -1,24 +1,0 @@
-import { Tag } from "@navikt/ds-react";
-import { withDsExample } from "components/website-modules/examples/withDsExample";
-
-const Example = () => {
-  return (
-    <div className="flex gap-2">
-      <Tag variant="alt3">Alt3</Tag>
-      <Tag variant="alt3-filled">Alt3</Tag>
-      <Tag variant="alt3-moderate">Alt3</Tag>
-    </div>
-  );
-};
-
-export default withDsExample(Example);
-
-/* Storybook story */
-export const Demo = {
-  render: Example,
-};
-
-export const args = {
-  index: 7,
-  desc: "Har ingen global betydning. Tjenesten/produktet kan bestemme hva de symboliserer.",
-};
