+++ aksel.nav.no/website/pages/eksempler/tag/default.tsx
@@ -3,10 +3,15 @@ import { withDsExample } from "components/website-modules/examples/withDsExample
 
 const Example = () => {
   return (
-    <div className="flex gap-2">
+    <div className="flex flex-wrap gap-2">
+      <Tag variant="neutral">Neutral</Tag>
       <Tag variant="info">Info</Tag>
-      <Tag variant="info-filled">Info</Tag>
-      <Tag variant="info-moderate">Info</Tag>
+      <Tag variant="success">Success</Tag>
+      <Tag variant="warning">Warning</Tag>
+      <Tag variant="error">Error</Tag>
+      <Tag variant="alt1">Alt1</Tag>
+      <Tag variant="alt2">Alt2</Tag>
+      <Tag variant="alt3">Alt3</Tag>
     </div>
   );
 };
@@ -19,6 +24,5 @@ export const Demo = {
 };
 
 export const args = {
-  index: 1,
-  desc: "Beskriver noe som er til opplysning, nyhet, etc.",
+  index: 0,
 };
