+++ aksel.nav.no/website/pages/eksempler/tag/filled.tsx
@@ -3,10 +3,15 @@ import { withDsExample } from "components/website-modules/examples/withDsExample
 
 const Example = () => {
   return (
-    <div className="flex gap-2">
-      <Tag variant="success">Success</Tag>
+    <div className="flex flex-wrap gap-2">
+      <Tag variant="neutral-filled">Neutral</Tag>
+      <Tag variant="info-filled">Info</Tag>
       <Tag variant="success-filled">Success</Tag>
-      <Tag variant="success-moderate">Success</Tag>
+      <Tag variant="warning-filled">Warning</Tag>
+      <Tag variant="error-filled">Error</Tag>
+      <Tag variant="alt1-filled">Alt1</Tag>
+      <Tag variant="alt2-filled">Alt2</Tag>
+      <Tag variant="alt3-filled">Alt3</Tag>
     </div>
   );
 };
@@ -20,5 +25,4 @@ export const Demo = {
 
 export const args = {
   index: 2,
-  desc: "Beskriver noe som er positivt, gjennomført, vellykket, etc.",
 };
