+++ aksel.nav.no/website/pages/eksempler/tag/small.tsx
@@ -3,9 +3,17 @@ import { withDsExample } from "components/website-modules/examples/withDsExample
 
 const Example = () => {
   return (
-    <Tag variant="info" size="small">
-      Info
-    </Tag>
+    <div className="flex flex-wrap gap-2">
+      <Tag size="small" variant="neutral">
+        Small
+      </Tag>
+      <Tag size="small" variant="neutral-moderate">
+        Small
+      </Tag>
+      <Tag size="small" variant="neutral-filled">
+        Small
+      </Tag>
+    </div>
   );
 };
 
@@ -17,5 +25,5 @@ export const Demo = {
 };
 
 export const args = {
-  index: 20,
+  index: 3,
 };
