--- aksel.nav.no/website/pages/eksempler/tag/warning.tsx
@@ -1,24 +1,0 @@
-import { Tag } from "@navikt/ds-react";
-import { withDsExample } from "components/website-modules/examples/withDsExample";
-
-const Example = () => {
-  return (
-    <div className="flex gap-2">
-      <Tag variant="warning">Warning</Tag>
-      <Tag variant="warning-filled">Warning</Tag>
-      <Tag variant="warning-moderate">Warning</Tag>
-    </div>
-  );
-};
-
-export default withDsExample(Example);
-
-/* Storybook story */
-export const Demo = {
-  render: Example,
-};
-
-export const args = {
-  index: 3,
-  desc: "Beskriver noe som er en advarsel, usikkert, noe som må utbedres, etc.",
-};
